Choose to use bower instead of xstatic packages
We want to use static libs as in thinkster-django-angular-boilerplate, but
the xstatic system does not have all of that. Therefore, we copied all
static libs management stuff from thinkster-django-angular-boilerplate
(with proper modifications). In particular, we will use bower as the
package manager for the static libs.

The xstatic code is merely disabled; its removal shall only be done when
we are sure.

The dependency stuff in bower.json is created by:
bower install -ddd angular angular-route angular-cookies bootstrap \
  bootstrap-material-design jquery ngDialog underscore --save