Introduce account and auth system
Basically, the default auth system is enabled and the web browser UI
login/logout works. Majority of the authentication app is copied from
the thinkster-django-angular-boilerplate project, with one noticable
exception: we use the default User model. This is to follow the guide
from the Django Design Patterns book of not doing so: we keep maximum
compatibility in everyway. Anyway, whenever there is some need for
extension, one-to-one relationship serves our purposes well.

Several key points:
0) Login via email or username are both OK.
1) Front-end development may fool browsers in unexpected manners. It is
wise to always use a clean profile dir for every new debug/test session.
2) Registering new users on web UI is currently not supported.
3) The 'django.contrib.auth.context_processors.auth' context processor
is required to have 'user' available in template rendering.